python_scraper_amazon/requirements.txt
# Amazon Affiliate Scraper - Python Dependencies

# Web scraping and automation
playwright==1.51.0
beautifulsoup4==4.12.3
lxml==5.3.0
selenium==4.27.1

# Database
psycopg2-binary==2.9.10

# Configuration and utilities
python-dotenv==1.0.1
pyyaml==6.0.2

# HTTP requests
requests==2.32.5

# Logging and monitoring
colorlog==6.9.0
